+++ weather@mockturtl/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Minor style changes
 * Make "Location Label Override" behavior identical to Panel label override.
+* Add option to override tooltip label for applet
 * Try to make hourly weather elements slimmer by putting precipitation change and volume in different rows
 * Update luxon and suncalc